% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whiteboard.R
\name{prompt}
\alias{prompt}
\title{Use prompts to update the whiteboard state}
\usage{
prompt(backdate = NULL)
}
\arguments{
\item{backdate}{Days to backdate (if NULL, prompt will appear)}
}
\value{
Invisibly returns the whiteboard tibble
}
\description{
Use prompts to update the whiteboard state
}
