% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whiteboard.R
\name{read}
\alias{read}
\title{Read the whiteboard data}
\usage{
read(file = getOption("whiteboard.data"))
}
\arguments{
\item{file}{Path to the csv file}
}
\value{
A tibble
}
\description{
Read the whiteboard data
}
