% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whiteboard.R
\name{update}
\alias{update}
\title{Update the whiteboard state}
\usage{
update(
  date = backdate(0),
  make_bed = TRUE,
  clean_up = TRUE,
  exercise = TRUE,
  no_alcohol = TRUE,
  no_cigarettes = TRUE,
  eat_okay = TRUE
)
}
\arguments{
\item{date}{Date}

\item{make_bed}{Logical}

\item{clean_up}{Logical}

\item{exercise}{Logical}

\item{no_alcohol}{Logical}

\item{no_cigarettes}{Logical}

\item{eat_okay}{Logical}
}
\value{
Invisibly returns the whiteboard tibble
}
\description{
Update the whiteboard state
}
