% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whiteboard.R
\name{write}
\alias{write}
\title{Write whiteboard data to file}
\usage{
write(dat, file = getOption("whiteboard.data"))
}
\arguments{
\item{dat}{Data frame}

\item{file}{File location}
}
\description{
Write whiteboard data to file
}
